--- a/internal/api/api.go
+++ b/internal/api/api.go
@@ -7,7 +7,6 @@ package api
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"net"
 	"net/http"
@@ -15,6 +14,7 @@ import (
 	"time"
 
 	"github.com/gorilla/mux"
+	jsoniter "github.com/json-iterator/go"
 	"github.com/spf13/viper"
 
 	"github.com/vmware-tanzu/octant/internal/config"
@@ -24,6 +24,8 @@ import (
 )
 
 //go:generate mockgen -destination=./fake/mock_service.go -package=fake github.com/vmware-tanzu/octant/internal/api Service
+
+var json = jsoniter.ConfigFastest
 
 const (
 	// ListenerAddrKey is the environment variable for the Octant listener address.