--- a/internal/api/breadcrumb_test.go
+++ b/internal/api/breadcrumb_test.go
@@ -6,7 +6,6 @@
 package api_test
 
 import (
-	"encoding/json"
 	"io/ioutil"
 	"path/filepath"
 	"testing"