--- a/internal/api/content_manager.go
+++ b/internal/api/content_manager.go
@@ -8,7 +8,6 @@ package api
 import (
 	"context"
 	"crypto/sha256"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"hash/fnv"