--- a/internal/api/namespaces_manager.go
+++ b/internal/api/namespaces_manager.go
@@ -8,7 +8,6 @@ package api
 import (
 	"bytes"
 	"context"
-	"encoding/json"
 
 	"github.com/pkg/errors"