--- a/internal/api/navigation_manager.go
+++ b/internal/api/navigation_manager.go
@@ -8,7 +8,6 @@ package api
 import (
 	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
 	"sync"