--- a/internal/api/websocket_client.go
+++ b/internal/api/websocket_client.go
@@ -7,7 +7,6 @@ package api
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"sync/atomic"
 	"time"