--- a/internal/octant/container_editor.go
+++ b/internal/octant/container_editor.go
@@ -7,7 +7,6 @@ package octant
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 
 	"github.com/pkg/errors"