--- a/internal/octant/object.go
+++ b/internal/octant/object.go
@@ -1,1 +1,5 @@
 package octant
+
+import jsoniter "github.com/json-iterator/go"
+
+var json = jsoniter.ConfigFastest