--- a/internal/printer/networkpolicy.go
+++ b/internal/printer/networkpolicy.go
@@ -7,7 +7,6 @@ package printer
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"strings"