--- a/internal/printer/volume.go
+++ b/internal/printer/volume.go
@@ -6,8 +6,6 @@ SPDX-License-Identifier: Apache-2.0
 package printer
 
 import (
-	"encoding/json"
-
 	corev1 "k8s.io/api/core/v1"
 
 	"github.com/vmware-tanzu/octant/pkg/view/component"