--- a/pkg/plugin/javascript.go
+++ b/pkg/plugin/javascript.go
@@ -7,7 +7,6 @@ package plugin
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"path"