--- a/pkg/plugin/javascript/httpclient.go
+++ b/pkg/plugin/javascript/httpclient.go
@@ -7,7 +7,6 @@ package javascript
 
 import (
 	"bytes"
-	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"net/http"