--- a/pkg/plugin/plugin.go
+++ b/pkg/plugin/plugin.go
@@ -9,10 +9,13 @@ import (
 	"context"
 
 	"github.com/hashicorp/go-plugin"
+	jsoniter "github.com/json-iterator/go"
 	"google.golang.org/grpc"
 
 	"github.com/vmware-tanzu/octant/pkg/plugin/dashboard"
 )
+
+var json = jsoniter.ConfigFastest
 
 var (
 	pluginMap = map[string]plugin.Plugin{