--- a/pkg/view/component/button.go
+++ b/pkg/view/component/button.go
@@ -1,8 +1,6 @@
 package component
 
 import (
-	"encoding/json"
-
 	"github.com/pkg/errors"
 
 	"github.com/vmware-tanzu/octant/pkg/action"