--- a/pkg/view/component/button_test.go
+++ b/pkg/view/component/button_test.go
@@ -1,7 +1,6 @@
 package component
 
 import (
-	"encoding/json"
 	"io/ioutil"
 	"path/filepath"
 	"testing"