--- a/pkg/view/component/code.go
+++ b/pkg/view/component/code.go
@@ -5,8 +5,6 @@ SPDX-License-Identifier: Apache-2.0
 
 package component
 
-import "encoding/json"
-
 // Value is a component for code
 //
 // +octant:component