--- a/pkg/view/component/code_test.go
+++ b/pkg/view/component/code_test.go
@@ -6,7 +6,6 @@ SPDX-License-Identifier: Apache-2.0
 package component
 
 import (
-	"encoding/json"
 	"testing"
 
 	"github.com/stretchr/testify/assert"