--- a/pkg/view/component/donut_chart.go
+++ b/pkg/view/component/donut_chart.go
@@ -5,8 +5,6 @@
 
 package component
 
-import "encoding/json"
-
 type DonutChartSize int
 
 const (