--- a/pkg/view/component/extension.go
+++ b/pkg/view/component/extension.go
@@ -1,8 +1,6 @@
 package component
 
 import (
-	"encoding/json"
-
 	"github.com/vmware-tanzu/octant/pkg/action"
 )