--- a/pkg/view/component/form_test.go
+++ b/pkg/view/component/form_test.go
@@ -1,7 +1,6 @@
 package component
 
 import (
-	"encoding/json"
 	"testing"
 
 	"github.com/stretchr/testify/assert"