--- a/pkg/view/component/grid_actions.go
+++ b/pkg/view/component/grid_actions.go
@@ -7,8 +7,6 @@
 package component
 
 import (
-	"encoding/json"
-
 	"github.com/vmware-tanzu/octant/pkg/action"
 )