--- a/pkg/view/component/labels.go
+++ b/pkg/view/component/labels.go
@@ -5,8 +5,6 @@ SPDX-License-Identifier: Apache-2.0
 
 package component
 
-import "encoding/json"
-
 var labelsFilteredKeys = []string{
 	"controller-revision-hash",
 	"controller-uid",