--- a/pkg/view/component/link.go
+++ b/pkg/view/component/link.go
@@ -5,8 +5,6 @@ SPDX-License-Identifier: Apache-2.0
 
 package component
 
-import "encoding/json"
-
 // Link is a text component that contains a link.
 //
 // +octant:component