--- a/pkg/view/component/list.go
+++ b/pkg/view/component/list.go
@@ -5,10 +5,6 @@ SPDX-License-Identifier: Apache-2.0
 
 package component
 
-import (
-	"encoding/json"
-)
-
 // List contains other Components
 //
 // +octant:component