--- a/pkg/view/component/modal.go
+++ b/pkg/view/component/modal.go
@@ -1,9 +1,5 @@
 package component
 
-import (
-	"encoding/json"
-)
-
 type ModalSize string
 
 const (