--- a/pkg/view/component/pod_status.go
+++ b/pkg/view/component/pod_status.go
@@ -5,8 +5,6 @@ SPDX-License-Identifier: Apache-2.0
 
 package component
 
-import "encoding/json"
-
 // PodSummary is a status summary for a pod.
 type PodSummary struct {
 	Details []Component `json:"details,omitempty"`