--- a/pkg/view/component/pod_status_test.go
+++ b/pkg/view/component/pod_status_test.go
@@ -6,7 +6,6 @@ SPDX-License-Identifier: Apache-2.0
 package component
 
 import (
-	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"path/filepath"