--- a/pkg/view/component/quadrant.go
+++ b/pkg/view/component/quadrant.go
@@ -6,7 +6,6 @@ SPDX-License-Identifier: Apache-2.0
 package component
 
 import (
-	"encoding/json"
 	"fmt"
 )