--- a/pkg/view/component/resource_viewer_test.go
+++ b/pkg/view/component/resource_viewer_test.go
@@ -6,7 +6,6 @@ SPDX-License-Identifier: Apache-2.0
 package component
 
 import (
-	"encoding/json"
 	"io/ioutil"
 	"path"
 	"testing"