--- a/pkg/view/component/selectors_test.go
+++ b/pkg/view/component/selectors_test.go
@@ -6,7 +6,6 @@ SPDX-License-Identifier: Apache-2.0
 package component
 
 import (
-	"encoding/json"
 	"io/ioutil"
 	"path"
 	"testing"