--- a/pkg/view/component/single_stat.go
+++ b/pkg/view/component/single_stat.go
@@ -5,8 +5,6 @@
 
 package component
 
-import "encoding/json"
-
 type SingleStatValue struct {
 	Text  string `json:"text"`
 	Color string `json:"color"`