--- a/pkg/view/component/stepper.go
+++ b/pkg/view/component/stepper.go
@@ -5,8 +5,6 @@ SPDX-License-Identifier: Apache-2.0
 
 package component
 
-import "encoding/json"
-
 // Stepper component implements json.Marshaler
 //
 // +octant:component