--- a/pkg/view/component/stepper_test.go
+++ b/pkg/view/component/stepper_test.go
@@ -1,7 +1,6 @@
 package component
 
 import (
-	"encoding/json"
 	"io/ioutil"
 	"path"
 	"testing"