--- a/pkg/view/component/summary.go
+++ b/pkg/view/component/summary.go
@@ -6,8 +6,6 @@ SPDX-License-Identifier: Apache-2.0
 package component
 
 import (
-	"encoding/json"
-
 	"github.com/vmware-tanzu/octant/internal/util/strings"
 )