--- a/pkg/view/component/table.go
+++ b/pkg/view/component/table.go
@@ -6,7 +6,6 @@ SPDX-License-Identifier: Apache-2.0
 package component
 
 import (
-	"encoding/json"
 	"fmt"
 	"sort"
 	"sync"