--- a/pkg/view/component/terminal.go
+++ b/pkg/view/component/terminal.go
@@ -6,7 +6,6 @@ SPDX-License-Identifier: Apache-2.0
 package component
 
 import (
-	"encoding/json"
 	"time"
 )