--- a/pkg/view/component/testing.go
+++ b/pkg/view/component/testing.go
@@ -1,7 +1,6 @@
 package component
 
 import (
-	"encoding/json"
 	"testing"
 
 	"github.com/stretchr/testify/assert"