--- a/pkg/view/component/timestamp.go
+++ b/pkg/view/component/timestamp.go
@@ -6,7 +6,6 @@ SPDX-License-Identifier: Apache-2.0
 package component
 
 import (
-	"encoding/json"
 	"time"
 )