… go.sum
… vendor/github.com/fatih/color/.travis.yml
… vendor/github.com/fatih/color/Gopkg.lock
… vendor/github.com/fatih/color/Gopkg.toml
… vendor/github.com/fatih/color/README.md
… vendor/github.com/fatih/color/go.mod
… vendor/github.com/fatih/color/go.sum
… vendor/github.com/iancoleman/strcase/.travis.yml
… vendor/github.com/iancoleman/strcase/acronyms.go
… vendor/github.com/iancoleman/strcase/camel.go
… vendor/github.com/iancoleman/strcase/numbers.go
… vendor/github.com/iancoleman/strcase/snake.go
… vendor/github.com/mattn/go-colorable/.travis.yml
… vendor/github.com/mattn/go-colorable/README.md
… vendor/github.com/mattn/go-colorable/colorable_appengine.go
… vendor/github.com/mattn/go-colorable/colorable_others.go
… vendor/github.com/mattn/go-colorable/colorable_windows.go
… vendor/github.com/mattn/go-colorable/go.mod
… vendor/github.com/mattn/go-colorable/go.sum
… vendor/github.com/mattn/go-colorable/go.test.sh
… vendor/github.com/mattn/go-isatty/.travis.yml
… vendor/github.com/mattn/go-isatty/README.md
… vendor/github.com/mattn/go-isatty/go.mod
… vendor/github.com/mattn/go-isatty/go.sum
… vendor/github.com/mattn/go-isatty/go.test.sh
… vendor/github.com/mattn/go-isatty/isatty_android.go
… vendor/github.com/mattn/go-isatty/isatty_bsd.go
… vendor/github.com/mattn/go-isatty/isatty_plan9.go
… vendor/github.com/mattn/go-isatty/isatty_tcgets.go
… vendor/github.com/mattn/go-isatty/renovate.json