@GuessWhoSamFoo made a great point in the last community meeting that we needed to develop a strategy to address CPU issues we are seeing in Octant. I already had this branch in flight and in my testing I'm seeing good enough improvements generally to make the external dependency worth it.

## Note
This is not an attempt to fix the high CPU spikes that have been reported on macOS. This was more a general exploration that I didn't want to die on the vine.

I'm also open to not landing this, but the way I see it, https://github.com/json-iterator/go is a standard library compatible well maintained JSON library that gives us speed improvements with zero effort.

These tests were captured against the same namespace, cluster, and conditions for each binary and the same actions were performed.

### Before *encoding/json* (grabbing the middle avg)
![image](https://user-images.githubusercontent.com/27317/110135750-e0833f00-7d9c-11eb-8424-03ac7cb2dc12.png)

### After *jsoniter* (grabbing the middle avg)
![image](https://user-images.githubusercontent.com/27317/110163960-b93d6980-7dbe-11eb-8daa-a5ac93939da5.png)


Signed-off-by: Wayne Witzel III <wayne@riotousliving.com>