@wwitzel3 2021-03-05
Switched to `ConfigFastest` and all of our Marshal/Unmarshal tests are passing with it and it is a notable speed improvement over `ConfigCompatibleWithStandardLibrary`