@GuessWhoSamFoo 2021-03-07
Thanks for making the swap. jsoniter (is it pronounced json-iter or jsoniter?) seems to get the performance boost by avoiding reflection although I'm unsure of the implications of this.

Ran these as a benchmark on as a unit test on the component level using table as an example:

```
func Benchmark_TableMarshal(b *testing.B) {
	tests := []struct {
		name         string
		input        Component
		expectedPath string
		isErr        bool
	}{
		{
			name: "general",
			input: &Table{
				Base: newBase(TypeTable, TitleFromString("my table")),
				Config: TableConfig{
					Filters: map[string]TableFilter{},
					Columns: []TableCol{
						{Name: "Name", Accessor: "Name"},
						{Name: "Description", Accessor: "Description"},
					},
					Rows: []TableRow{
						{
							"Name": &Text{
								Config: TextConfig{
									Text: "First",
								},
							},
							"Description": &Text{
								Config: TextConfig{
									Text: "The first row",
								},
							},
						},
						{
							"Name": &Text{
								Config: TextConfig{
									Text: "Last",
								},
							},
							"Description": &Text{
								Config: TextConfig{
									Text: "The last row",
								},
							},
						},
					},
					EmptyContent: "",
					Loading:      false,
				},
			},
			expectedPath: "table.json",
		},
	}

	for _, tc := range tests {
		b.Run(tc.name, func(b *testing.B) {
			b.ReportAllocs()
			actual, err := json.Marshal(tc.input)
			isErr := err != nil
			if isErr != tc.isErr {
				b.Fatalf("Unexpected error: %v", err)
			}

			expected, err := ioutil.ReadFile(filepath.Join("testdata", tc.expectedPath))
			require.NoError(b, err, "reading test fixtures")
			assert.JSONEq(b, string(expected), string(actual))
		})
	}
}
```

This PR:
```
goos: linux
goarch: amd64
pkg: github.com/vmware-tanzu/octant/pkg/view/component
cpu: AMD Ryzen 7 5800X 8-Core Processor             
Benchmark_TableMarshal/general-16         	1000000000	         0.0002186 ns/op	       0 B/op	       0 allocs/op
Benchmark_TableMarshal/general-16         	1000000000	         0.0002111 ns/op	       0 B/op	       0 allocs/op
Benchmark_TableMarshal/general-16         	1000000000	         0.0001815 ns/op	       0 B/op	       0 allocs/op
Benchmark_TableMarshal/general-16         	1000000000	         0.0001779 ns/op	       0 B/op	       0 allocs/op
Benchmark_TableMarshal/general-16         	1000000000	         0.0001707 ns/op	       0 B/op	       0 allocs/op
Benchmark_TableMarshal/general-16         	1000000000	         0.0001628 ns/op	       0 B/op	       0 allocs/op
Benchmark_TableMarshal/general-16         	1000000000	         0.0001679 ns/op	       0 B/op	       0 allocs/op
Benchmark_TableMarshal/general-16         	1000000000	         0.0001756 ns/op	       0 B/op	       0 allocs/op
Benchmark_TableMarshal/general-16         	1000000000	         0.0001785 ns/op	       0 B/op	       0 allocs/op
Benchmark_TableMarshal/general-16         	1000000000	         0.0000918 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/vmware-tanzu/octant/pkg/view/component	0.160s
```

Main:
``` 
goos: linux
goarch: amd64
pkg: github.com/vmware-tanzu/octant/pkg/view/component
cpu: AMD Ryzen 7 5800X 8-Core Processor             
Benchmark_TableMarshal/general-16         	1000000000	         0.0001699 ns/op	       0 B/op	       0 allocs/op
Benchmark_TableMarshal/general-16         	1000000000	         0.0001631 ns/op	       0 B/op	       0 allocs/op
Benchmark_TableMarshal/general-16         	1000000000	         0.0001777 ns/op	       0 B/op	       0 allocs/op
Benchmark_TableMarshal/general-16         	1000000000	         0.0001628 ns/op	       0 B/op	       0 allocs/op
Benchmark_TableMarshal/general-16         	1000000000	         0.0001853 ns/op	       0 B/op	       0 allocs/op
Benchmark_TableMarshal/general-16         	1000000000	         0.0001321 ns/op	       0 B/op	       0 allocs/op
Benchmark_TableMarshal/general-16         	1000000000	         0.0001838 ns/op	       0 B/op	       0 allocs/op
Benchmark_TableMarshal/general-16         	1000000000	         0.0001732 ns/op	       0 B/op	       0 allocs/op
Benchmark_TableMarshal/general-16         	1000000000	         0.0000840 ns/op	       0 B/op	       0 allocs/op
Benchmark_TableMarshal/general-16         	1000000000	         0.0001779 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/vmware-tanzu/octant/pkg/view/component	0.138s
```

While there is still a performance increase, it is not as drastic as shown above. We might observe this effect better through benchmarking nested components or the entire content response.