@mklanjsek 2021-03-08
Tested this branch on my mac laptop and I see 10-15% performance boost 🎉